requirements.txt
numpy>=1.21.0
opencv-python-headless>=4.5.0
mediapipe>=0.9.0
Pillow>=9.0.0
tqdm>=4.62.0
cog>=0.8.0
